feat(chat): open user profile sheet on tapping sender avatar
Extract the room member profile sheet into a public MemberSheetDialog
plus a showMemberSheet() helper, and reuse it from chat. Tapping a
sender avatar now opens the same sheet showing avatar, display name,
Matrix ID, presence and role, with copy-MXID, send-message and
moderation actions gated by permission.

Closes #669
